refactor: OCP 적용
