+++ Spring-MVC-Part1/servlet/src/main/java/hello/servlet/web/frontController/v5/FrontControllerServletV5.java
@@ -2,14 +2,7 @@ package hello.servlet.web.frontController.v5;
 
 import hello.servlet.web.frontController.ModelView;
 import hello.servlet.web.frontController.MyView;
-import hello.servlet.web.frontController.v3.controller.MemberFormControllerV3;
-import hello.servlet.web.frontController.v3.controller.MemberListControllerV3;
-import hello.servlet.web.frontController.v3.controller.MemberSaveControllerV3;
-import hello.servlet.web.frontController.v4.controller.MemberFormControllerV4;
-import hello.servlet.web.frontController.v4.controller.MemberListControllerV4;
-import hello.servlet.web.frontController.v4.controller.MemberSaveControllerV4;
-import hello.servlet.web.frontController.v5.adapter.ControllerV3HandlerAdapter;
-import hello.servlet.web.frontController.v5.adapter.ControllerV4HandlerAdapter;
+import hello.servlet.web.frontController.v5.adapter.HandlerAdapterFactory;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.annotation.WebServlet;
 import jakarta.servlet.http.HttpServlet;
@@ -17,67 +10,30 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 
 import java.io.IOException;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
 
 @WebServlet(name = "frontControllerServletV5", urlPatterns = "/front-controller/v5/*")
 public class FrontControllerServletV5 extends HttpServlet {
 
-    private final Map<String, Object> handlerMappingMap = new HashMap<>();
-    private final List<MyHandlerAdapter> handlerAdapters = new ArrayList<>();
+    private final HandlerAdapterFactory handlerAdapterFactory;
 
     public FrontControllerServletV5() {
-        initHandlerMappingMap();
-        initHandlerAdapters();
+        handlerAdapterFactory = new HandlerAdapterFactory();
     }
 
     @Override
     protected void service(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        Object handler = getHandler(request);
+        try {
+            Object handler = handlerAdapterFactory.getHandler(request.getRequestURI());
+            MyHandlerAdapter adapter = handlerAdapterFactory.getAdapter(handler);
 
-        if (handler == null) {
+            ModelView mv = adapter.handle(request, response, handler);
+            MyView view = viewResolver(mv.getViewName());
+
+            view.render(mv.getModel(), request, response);
+        } catch (IllegalArgumentException e) {
+            System.out.println(e.getMessage());
             response.setStatus(HttpServletResponse.SC_NOT_FOUND);
-            return;
         }
-
-        MyHandlerAdapter adapter = getHandlerAdapter(handler);
-
-        ModelView mv = adapter.handle(request, response, handler);
-        MyView view = viewResolver(mv.getViewName());
-
-        view.render(mv.getModel(), request, response);
-    }
-
-    private void initHandlerMappingMap() {
-        handlerMappingMap.put("/front-controller/v5/v3/members/new-form", new MemberFormControllerV3());
-        handlerMappingMap.put("/front-controller/v5/v3/members/save", new MemberSaveControllerV3());
-        handlerMappingMap.put("/front-controller/v5/v3/members", new MemberListControllerV3());
-
-        handlerMappingMap.put("/front-controller/v5/v4/members/new-form", new MemberFormControllerV4());
-        handlerMappingMap.put("/front-controller/v5/v4/members/save", new MemberSaveControllerV4());
-        handlerMappingMap.put("/front-controller/v5/v4/members", new MemberListControllerV4());
-    }
-
-    private void initHandlerAdapters() {
-        handlerAdapters.add(new ControllerV3HandlerAdapter());
-        handlerAdapters.add(new ControllerV4HandlerAdapter());
-    }
-
-    private Object getHandler(HttpServletRequest request) {
-        String requestURI = request.getRequestURI();
-        return handlerMappingMap.get(requestURI);
-    }
-
-    private MyHandlerAdapter getHandlerAdapter(Object handler) {
-        for (MyHandlerAdapter adapter : handlerAdapters) {
-            if (adapter.supports(handler)) {
-                return adapter;
-            }
-        }
-
-        throw new IllegalArgumentException("Handler Adapter를 찾을 수 없습니다. handler: " + handler);
     }
 
     private MyView viewResolver(String viewName) {
