+++ Spring-MVC-Part1/servlet/src/main/java/hello/servlet/web/frontController/v5/adapter/HandlerAdapterFactory.java
@@ -1,0 +1,61 @@
+package hello.servlet.web.frontController.v5.adapter;
+
+import hello.servlet.web.frontController.v3.controller.MemberFormControllerV3;
+import hello.servlet.web.frontController.v3.controller.MemberListControllerV3;
+import hello.servlet.web.frontController.v3.controller.MemberSaveControllerV3;
+import hello.servlet.web.frontController.v4.controller.MemberFormControllerV4;
+import hello.servlet.web.frontController.v4.controller.MemberListControllerV4;
+import hello.servlet.web.frontController.v4.controller.MemberSaveControllerV4;
+import hello.servlet.web.frontController.v5.MyHandlerAdapter;
+
+import java.util.*;
+
+public class HandlerAdapterFactory {
+
+    private final Map<String, Object> handlerMappingMap = new HashMap<>();
+    private final List<MyHandlerAdapter> handlerAdapters = new ArrayList<>();
+
+    public HandlerAdapterFactory() {
+        initHandlerMappingMap();
+        initHandlerAdapters();
+    }
+
+    public Object getHandler(String requestURI) {
+        Object handler = handlerMappingMap.get(requestURI);
+
+        if (handler == null) {
+            throw new IllegalArgumentException("URI를 찾을 수 없습니다. URI: " + requestURI);
+        }
+
+        return handler;
+    }
+
+    public MyHandlerAdapter getAdapter(Object handler) {
+        return getHandlerAdapter(handler);
+    }
+
+    private void initHandlerMappingMap() {
+        handlerMappingMap.put("/front-controller/v5/v3/members/new-form", new MemberFormControllerV3());
+        handlerMappingMap.put("/front-controller/v5/v3/members/save", new MemberSaveControllerV3());
+        handlerMappingMap.put("/front-controller/v5/v3/members", new MemberListControllerV3());
+
+        handlerMappingMap.put("/front-controller/v5/v4/members/new-form", new MemberFormControllerV4());
+        handlerMappingMap.put("/front-controller/v5/v4/members/save", new MemberSaveControllerV4());
+        handlerMappingMap.put("/front-controller/v5/v4/members", new MemberListControllerV4());
+    }
+
+    private void initHandlerAdapters() {
+        handlerAdapters.add(new ControllerV3HandlerAdapter());
+        handlerAdapters.add(new ControllerV4HandlerAdapter());
+    }
+
+    private MyHandlerAdapter getHandlerAdapter(Object handler) {
+        for (MyHandlerAdapter adapter : handlerAdapters) {
+            if (adapter.supports(handler)) {
+                return adapter;
+            }
+        }
+
+        throw new IllegalArgumentException("Handler Adapter를 찾을 수 없습니다. handler: " + handler);
+    }
+}
